reverse_image_process/requirements.txt
torch
torchvision
transformers
Pillow
numpy
matplotlib